+++ node/src/test/test-PipeTransport.ts
@@ -274,12 +274,6 @@ test('router.pipeToRouter() succeeds with audio', async () => {
 			uri: 'https://aomediacodec.github.io/av1-rtp-spec/#dependency-descriptor-rtp-header-extension',
 		},
 		{
-			uri: 'http://www.webrtc.org/experiments/rtp-hdrext/abs-capture-time',
-			id: 10,
-			encrypt: false,
-			parameters: {},
-		},
-		{
 			uri: 'http://www.webrtc.org/experiments/rtp-hdrext/playout-delay',
 			id: 11,
 			encrypt: false,
@@ -332,12 +326,6 @@ test('router.pipeToRouter() succeeds with audio', async () => {
 			id: 7,
 			parameters: {},
 			uri: 'https://aomediacodec.github.io/av1-rtp-spec/#dependency-descriptor-rtp-header-extension',
-		},
-		{
-			uri: 'http://www.webrtc.org/experiments/rtp-hdrext/abs-capture-time',
-			id: 10,
-			encrypt: false,
-			parameters: {},
 		},
 		{
 			uri: 'http://www.webrtc.org/experiments/rtp-hdrext/playout-delay',
@@ -413,12 +401,6 @@ test('router.pipeToRouter() succeeds with video', async () => {
 			parameters: {},
 		},
 		{
-			uri: 'http://www.webrtc.org/experiments/rtp-hdrext/abs-capture-time',
-			id: 10,
-			encrypt: false,
-			parameters: {},
-		},
-		{
 			uri: 'http://www.webrtc.org/experiments/rtp-hdrext/playout-delay',
 			id: 11,
 			encrypt: false,
@@ -479,12 +461,6 @@ test('router.pipeToRouter() succeeds with video', async () => {
 			parameters: {},
 		},
 		{
-			uri: 'http://www.webrtc.org/experiments/rtp-hdrext/abs-capture-time',
-			id: 10,
-			encrypt: false,
-			parameters: {},
-		},
-		{
 			uri: 'http://www.webrtc.org/experiments/rtp-hdrext/playout-delay',
 			id: 11,
 			encrypt: false,
@@ -498,6 +474,47 @@ test('router.pipeToRouter() succeeds with video', async () => {
 		},
 	]);
 	expect(pipeProducer.paused).toBe(true);
+}, 2000);
+
+test('router.pipeToRouter() gives abs-capture-time to the pipe Consumer if the Producer has it', async () => {
+	const videoProducer2 = await ctx.webRtcTransport1!.produce({
+		kind: 'video',
+		rtpParameters: {
+			mid: 'VIDEO2',
+			codecs: [
+				{
+					mimeType: 'video/VP8',
+					payloadType: 112,
+					clockRate: 90000,
+				},
+			],
+			headerExtensions: [
+				{
+					uri: 'http://www.webrtc.org/experiments/rtp-hdrext/abs-capture-time',
+					id: 12,
+				},
+			],
+			encodings: [{ ssrc: 44444444 }],
+			rtcp: {
+				cname: 'FOOBAR',
+			},
+		},
+	});
+
+	const { pipeConsumer } = (await ctx.router1!.pipeToRouter({
+		producerId: videoProducer2.id,
+		router: ctx.router2!,
+	})) as {
+		pipeConsumer: mediasoup.types.Consumer;
+		pipeProducer: mediasoup.types.Producer;
+	};
+
+	expect(pipeConsumer.rtpParameters.headerExtensions).toContainEqual({
+		uri: 'http://www.webrtc.org/experiments/rtp-hdrext/abs-capture-time',
+		id: 10,
+		encrypt: false,
+		parameters: {},
+	});
 }, 2000);
 
 test('router.pipeToRouter() with unknown router, producerId or dataProducerId fails', async () => {
@@ -624,12 +641,6 @@ test('router.createPipeTransport() with enableRtx succeeds', async () => {
 		{
 			uri: 'urn:ietf:params:rtp-hdrext:toffset',
 			id: 9,
-			encrypt: false,
-			parameters: {},
-		},
-		{
-			uri: 'http://www.webrtc.org/experiments/rtp-hdrext/abs-capture-time',
-			id: 10,
 			encrypt: false,
 			parameters: {},
 		},